acp-agent-communication-protocol/pyproject.toml
[project]
name = "acp-agent-communication-protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[tools]>=0.141.0",
    "crewai-tools>=0.51.1",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.1.1",
    "opentelemetry-instrumentation-openai-v2>=2.1b0",
    "langfuse>=3.0.0",
    "openinference-instrumentation-crewai>=0.1.10",
    "openinference-instrumentation-litellm>=0.1.23",
    "acp-sdk>=1.0.0",
]
